Final Version
The Changes made were to ChatClient.java, ClientConsole.java and EchoServer.java.

ServerConsole.java was written by modifying ClientConsole.java to allow the server to interact with the command line.

All changes made to these files were made in order to add the functionnalities depicted in the layout for SEG2105 - assignment 1.