Complete OCSF phase 2 - with testcases
This is my complete implementation of OCSF phase 2.

I was unable to run the Test Cases from 2014, 2015, 2016, 2018 and 2019 simply
because they all require two computers and currently, there is only one computer
in my household.

Also, some of the messages displayed from certain connections/disconnections might
differ from what was intended in the assignment but I think you will still find
that it functions the same.